requirement.txt
fastapi
uvicorn
alembic
asyncpg
sqlalchemy
dotenv
sqlalchemy[asyncio]
python-jose[cryptography]
passlib
bcrypt==4.0.1
python-multipart